naming convention: now reads left-to-right as is normal for golang
(e.g., the word "map" or "list" now comes first.)

Separators are also clearer.

Previously: "KVmap", "Vlist".

Now: "map__k__v", "list__v".

I considered making this itself templatable, but balked due to the
finite hours in a day (and in particular, mostly because golang
convention would have the capitalization rules vary depending on
what's first and how close it is to other things, which sounded
like a pain to deal with).  PRs would be welcomed.